chore: enhance architecture
