+++ lib/features/projects/controllers/presentation_controller.dart
@@ -1,20 +1,11 @@
-// di bridge
-import 'package:datn_mobile/features/projects/data/providers/repository_provider.dart';
-import 'package:datn_mobile/features/projects/domain/entity/presentation.dart';
-import 'package:datn_mobile/features/projects/domain/entity/presentation_minimal.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+part of 'controller_provider.dart';
 
 // Query
-final presentationsControllerProvider =
-    AsyncNotifierProvider<PresentationsController, List<PresentationMinimal>>(
-      () => PresentationsController(),
-    );
-
 class PresentationsController extends AsyncNotifier<List<PresentationMinimal>> {
   @override
   Future<List<PresentationMinimal>> build() async {
     final response = await ref
-        .read(presentationRepositoryProvider)
+        .read(presentationServiceProvider)
         .fetchPresentations();
 
     return response;
@@ -23,18 +14,13 @@ class PresentationsController extends AsyncNotifier<List<PresentationMinimal>> {
   Future<void> refresh() async {
     state = const AsyncLoading();
     final response = await ref
-        .read(presentationRepositoryProvider)
+        .read(presentationServiceProvider)
         .fetchPresentations();
     state = await AsyncValue.guard(() => Future.value(response));
   }
 }
 
 // Command
-final createPresentationControllerProvider =
-    AsyncNotifierProvider<CreatePresentationController, void>(
-      () => CreatePresentationController(),
-    );
-
 class CreatePresentationController extends AsyncNotifier<void> {
   @override
   Future<void> build() async {}
@@ -42,9 +28,7 @@ class CreatePresentationController extends AsyncNotifier<void> {
   Future<void> create(Presentation presentation) async {
     state = const AsyncLoading();
     state = await AsyncValue.guard(() async {
-      await ref
-          .read(presentationRepositoryProvider)
-          .addPresentation(presentation);
+      await ref.read(presentationServiceProvider).addPresentation(presentation);
       await ref.read(presentationsControllerProvider.notifier).refresh();
     });
   }
