+++ lib/features/projects/data/repository/presentation_repository_impl.dart
@@ -1,12 +1,7 @@
-import 'package:datn_mobile/features/projects/data/dto/presentation_dto.dart';
-import 'package:datn_mobile/features/projects/data/dto/presentation_minimal_dto.dart';
-import 'package:datn_mobile/features/projects/data/source/resource_remote_source.dart';
-import 'package:datn_mobile/features/projects/domain/entity/presentation.dart';
-import 'package:datn_mobile/features/projects/domain/entity/presentation_minimal.dart';
-import 'package:datn_mobile/features/projects/domain/repository/presentation_repository.dart';
+part of 'repository_provider.dart';
 
 class PresentationRepositoryImpl implements PresentationRepository {
-  final ResourceRemoteSource _remoteSource;
+  final ProjectsRemoteSource _remoteSource;
   PresentationRepositoryImpl(this._remoteSource);
 
   @override
@@ -15,7 +10,6 @@ class PresentationRepositoryImpl implements PresentationRepository {
     // PresentationDto presentationDto = presentation.toDto();
     // LOGIC...
     _remoteSource.createPresentation(presentation.toDto());
-    // TODO: implement addBook
     throw UnimplementedError();
   }
